+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ai-sdk/google-vertex
 
+## 4.0.0-beta.20
+
+### Patch Changes
+
+- 95f65c2: chore: use import \* from zod/v4
+- Updated dependencies [95f65c2]
+- Updated dependencies [95f65c2]
+  - @ai-sdk/provider-utils@4.0.0-beta.11
+  - @ai-sdk/anthropic@3.0.0-beta.15
+  - @ai-sdk/google@3.0.0-beta.15
+
 ## 4.0.0-beta.19
 
 ### Patch Changes